+++ client/src/components/characters/character/stats/StatLevel.vue
